requirements.txt
django
daphne
channels
django-allauth
PyJWT
requests